requirements.txt
# Core Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Numerical Computing
numpy>=1.24.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Handling
pandas>=2.0.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Training Utilities
tensorboard>=2.13.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipywidgets>=8.0.0

# Optional: Advanced features
# albumentations>=1.3.0  # Advanced augmentation
# scikit-learn>=1.3.0    # Additional metrics
